Updated patch and delete actions for recipes and ingredients

+++ api/routes/ingredients.js
@@ -113,8 +113,8 @@ function authenticatedActions(router) {
     router.patch(constructPath(endpointPath, '/:id'), async function(req, res) {
         // Attempt to update ingredient
         Ingredient.findByIdAndUpdate(req.params.id, req.body)
-        .then(function() {
-            res.json('Ingredient updated');
+        .then(function(ingredient) {
+            res.json(ingredient);
         })
         .catch(function() {
             res.status(422).send("Ingredient update failed.");
@@ -123,15 +123,24 @@ function authenticatedActions(router) {
 
     // DELETE /:id, deletes an ingredient by id
     router.delete(constructPath(endpointPath, '/:id'), async function(req, res) {
+        const token = req.headers.authorization.split(' ')[1];
+        const { userId } = jwt.verifyJWT(token);
+
+        // Check if the user ids match
+        if (userId != req.params.id) {
+            res.status(422).json({ error: "The User IDs do not match."});
+            return;
+        }
+
         // Attempt to delete ingredient
         Ingredient.findById(req.params.id, async function(err, ingredient) {
             if (!ingredient) {
                 res.status(404).send('Ingredient not found');
             } else {
                 Ingredient.findByIdAndRemove(req.params.id)
-                .then(function() {res.status(200).json("Article deleted") })
+                .then(function() {res.status(200).json("Ingredient deleted.") })
                 .catch(function(err) {
-                    res.status(400).send("Article delete failed.");
+                    res.status(400).send("Ingredient delete failed.");
                 })
             }
         });
